% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EBICglassoNet.R
\name{EBICglassoNet}
\alias{EBICglassoNet}
\title{Gaussian Markov random field estimation using graphical LASSO and extended Bayesian information criterion}
\usage{
EBICglassoNet(data, gamma = NULL, ...)
}
\arguments{
\item{data}{a data frame, each column presents a node, there should be no miss values in the data frame.}

\item{gamma}{EBIC hyperparameter in [0,1]. NULL selects 0.5.}

\item{...}{Reserved for future extensions.}
}
\description{
Estimate an EBICglasso network and return a quicknet_fit object.
}
\examples{
data('mtcars')
EBICglassoNet(mtcars)

}
