% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_edges_df.R
\name{get_edges_df}
\alias{get_edges_df}
\title{Get data frame of edges and weights of one or two networks}
\usage{
get_edges_df(net1, net2 = NULL, method = "union", labels = NULL)
}
\arguments{
\item{net1}{a network. Should be either the product of \code{qgraph} or \code{quickNet}.}

\item{net2}{alternative. a network. Should be either the product of \code{qgraph} or \code{quickNet}.}

\item{method}{only works when net2 is provided. \itemize{
\item \code{"union"}: return the edges that exist in net1 or net2, default.
\item \code{"intersect"}: return the edges that both exist in net1 and net2.
}}

\item{labels}{the name of each node. If provided, the nodes will be named by labels.}
}
\value{
a data frame with 3 columns. The vectors in from and to represent the node index.
}
\description{
Get data frame of edges and weights of one or two networks
}
\examples{

data("mtcars")
net1 <- quickNet(mtcars)
net2 <- quickNet(mtcars^3)
edges1 <- get_edges_df(net1)
edges2 <- get_edges_df(net1, net2, method = 'intersect')

}
