% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_strength_node_size.R
\name{get_strength_node_size}
\alias{get_strength_node_size}
\title{Get new node size for graph based on strength centrality}
\usage{
get_strength_node_size(Centrality, scale_add = 0.5, scale_times = 2.5)
}
\arguments{
\item{Centrality}{The output of function \code{Centrality}}

\item{scale_add}{The basic size added to original size, in case of the node is too small}

\item{scale_times}{The scale times of the original size in case the node is too small}
}
\value{
a named vector containing the new sizes for each node.
}
\description{
Get new node size for graph based on strength centrality
}
\examples{
data("mtcars")
network = quickNet(mtcars)
centrality = Centrality(network)
vsize = get_strength_node_size(centrality)

quickNet(mtcars, vsize = vsize)
}
