% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalCoeff.R
\name{globalCoeff}
\alias{globalCoeff}
\title{Compute global coefficients of a network}
\usage{
globalCoeff(x, list = FALSE)
}
\arguments{
\item{x}{a network produced from \code{quickNet}, or an adjacency matrix of network data.}

\item{list}{whether arrange the results as a list. The default is FALSE.}
}
\value{
\itemize{
\item \code{globalStrength}: the global strength.
\item \code{AGS}: the average of all the edge weights in the network. It
equals globalStrength/(N * (N-1)/2) for undirected networks, or
globalStrength/(N * (N-1)) for directed networks. Self loops are excluded.
\item \code{ASPL}: the global average shortest path length.
\item \code{CC}: the global clustering coefficient, averaged across nodes.
}
}
\description{
Compute global coefficients of a network
}
\examples{

data('mtcars')

globalCoeff <- globalCoeff(quickNet(mtcars))

}
