% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_validation.R
\name{plot.quicknet_power_validation}
\alias{plot.quicknet_power_validation}
\title{Plot native powerly validation results}
\usage{
\method{plot}{quicknet_power_validation}(x, ...)
}
\arguments{
\item{x}{A quicknet_power_validation object.}

\item{...}{Arguments forwarded to powerly's native validation plot method.}
}
\value{
The native plot result.
}
\description{
Plot native powerly validation results
}
