% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imports.R
\docType{package}
\name{quickNet-package}
\alias{quickNet-package}
\title{quickNet: Quick Psychological Network Construction}
\description{
Provides a quick interface for estimating, visualizing, and comparing psychological networks.
}
\author{
\strong{Maintainer}: Lei Guo \email{guolei0812@yeah.net}

}
