% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bridge.R
\name{Bridge}
\alias{Bridge}
\title{Title Calculate Bridge Coefficients}
\usage{
Bridge(
  net_G,
  communities = NULL,
  useCommunities = "all",
  include = "basic",
  normalize = T,
  ...
)
}
\arguments{
\item{net_G}{a qgraph object, could also be the result of quickNet.}

\item{communities}{an object of class "communities" (igraph) OR a character vector of community assignments for each node (e.g., c("Comm1", "Comm1", "Comm2", "Comm2)). The ordering of this vector should correspond to the vector from argument "nodes". Can also be in list format (e.g., list("Comm1"=c(1:10), "Comm2"=c(11:20))).}

\item{useCommunities}{character vector specifying which communities should be included. Default set to "all".}

\item{include}{the statistics calculated. Default "basic" calculates Bridge strength, Bridge betweenness and Bridge closeness. if "all" or "All", the Bridge expected influence (1-step) and Bridge expected influence (2-step) will also be calculated. You can also specify the statistics you want. See \code{networktools::bridge}.}

\item{normalize}{logical. Bridge centralities are divided by their highest possible value (assuming max edge strength=1) in order to normalize by different community sizes.}

\item{...}{other parameters from \code{networktools::bridge}}
}
\value{
a list contains the bridge coefficient information:\itemize{
\item\code{bridgePlot:} the plot of bridge coefficients of each node.
\item\code{bridge_data:} the result of \code{networktools::bridge}.}
}
\description{
Title Calculate Bridge Coefficients
}
\examples{
data('mtcars')
Bridge <- Bridge(quickNet(mtcars), communities = list(c1 = 1:5, c2 = 6:11))

}
