% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bridgeGroup.R
\name{bridgeGroup}
\alias{bridgeGroup}
\title{Get new groups that bridge nodes are in a single group.}
\usage{
bridgeGroup(bridge, groups, labels = NULL, probs = 0.8)
}
\arguments{
\item{bridge}{The output of \code{quickNet::Bridge}.}

\item{groups}{The original groups or communities parameter used.}

\item{labels}{If there are labels for each node, please provide.}

\item{probs}{The Bridge Strength threshold, 0~1. The nodes whose Bridge Strength bigger than \code{probs} will be assigned to a new group. Default is 0.8.}
}
\value{
A factor, can be used as new \code{groups} parameter in \code{quickNet::quickNet}
}
\description{
Get new groups that bridge nodes are in a single group.
}
\examples{
data('mtcars')
groups = list(A = 1:5, B = 6:11)
net_G <- quickNet(mtcars, groups = groups)
Bridge <- Bridge(net_G,communities = groups)
new_groups <- bridgeGroup(Bridge, groups)
net_G <- quickNet(mtcars, groups = new_groups)

}
