% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bridgeGroup.R
\name{bridgeGroup}
\alias{bridgeGroup}
\title{Get new groups that bridge nodes are in a single group.}
\usage{
bridgeGroup(bridge, groups, labels = NULL, n = 1, by_group = TRUE)
}
\arguments{
\item{bridge}{The output of \code{quickNet::Bridge}.}

\item{groups}{The original groups or communities parameter used. if groups is a list, the name of each group should not be end with numbers.}

\item{labels}{If there are labels for each node, please provide.}

\item{n}{The number of nodes to recode by bridge strength, \code{n} nodes with highest bridge strength will be recoded as bridge node.}

\item{by_group}{Whether the node recoding procedure should be conducted by groups. If TRUE, the function will recode n nodes with highest bridge strength in each original group. If FALSE,the function will recode n nodes with highest bridge strength in all nodes.}
}
\value{
A factor, can be used as new \code{groups} parameter in \code{quickNet::quickNet}
}
\description{
Get new groups that bridge nodes are in a single group.
}
\examples{
data('mtcars')
groups = list(A = 1:5, B = 6:11)
net_G <- quickNet(mtcars, groups = groups)
Bridge <- Bridge(net_G,communities = groups)
new_groups <- bridgeGroup(Bridge, groups, n = 1, by_group = TRUE)
net_G <- quickNet(mtcars, groups = new_groups)

}
