% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Centrality.R
\name{Centrality}
\alias{Centrality}
\title{estimate the node Strength, Betweenness and Closeness of a network}
\usage{
Centrality(network_G, include = "all", ...)
}
\arguments{
\item{network_G}{a qgraph object, could also be the result of quickNet.}

\item{include}{a vector of measures to include. if missing all measures available will be included. Not included by default are "Closeness", "Betweenness", "ExpectedInfluence", "OutExpectedInfluence", and "InExpectedInfluence". Can also be "all" or "All" to include all available centrality measures.}

\item{...}{other parameters from \code{qgraph::centralityPlot}, this only affect the output of Centrality$centralityPlot.}
}
\value{
a list contains the centrality information:\itemize{
\item\code{centralityPlot:} the result of \code{qgraph::centralityPlot}.
\item\code{centrailty_data:} the result of \code{qgraph::centrality}.}
}
\description{
estimate the node Strength, Betweenness and Closeness of a network
}
\examples{
data('mtcars')
Centrality <- Centrality(quickNet(mtcars))
}
