% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EBICglassoNet.R
\name{EBICglassoNet}
\alias{EBICglassoNet}
\title{Gaussian Markov random field estimation using graphical LASSO and extended Bayesian information criterion}
\usage{
EBICglassoNet(data, ...)
}
\arguments{
\item{data}{a data frame, each column presents a node, there should be no miss values in the data frame.}

\item{...}{paramaters from \code{bootnet::estimateNetwork}}
}
\description{
short version of botnet::estimateNetwork(data, default = 'EBICglasso')
}
\examples{
data('mtcars')
EBICglassoNet(mtcars)

}
