% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bridge_plot.R
\name{get_bridge_plot}
\alias{get_bridge_plot}
\title{get bridge plot from bridge results.}
\usage{
get_bridge_plot(
  bridge,
  prefix = "",
  path = ".",
  device = "pdf",
  width = 10,
  height = 7,
  get.table = TRUE,
  ...
)
}
\arguments{
\item{bridge}{output from \code{quickNet::Bridge}}

\item{prefix}{the prefix of output plot files.}

\item{path}{the path of output files, can be either a relative or absolute path.}

\item{device}{any format supported by \code{ggplot2::ggsave}.}

\item{width}{the width of plot, in inch.}

\item{height}{the height of plot, in inch.}

\item{get.table}{logical. whether get the csv file of bridge coefficient. Default is TRUE.}

\item{...}{other parameter from \code{ggplot2::ggsave}.}
}
\value{
a Bridge plot will be output to the specified path.
}
\description{
a wrapper function of \code{ggplot2::ggsave}
}
\examples{
data('mtcars')
Bridge <- Bridge(quickNet(mtcars), communities = c(rep(1,5),rep(2,6)))
get_bridge_plot(Bridge, prefix = 'test')

}
