% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_centrality_plot.R
\name{get_centrality_plot}
\alias{get_centrality_plot}
\title{get network plot from stability results.}
\usage{
get_centrality_plot(
  centrality,
  prefix = "",
  path = ".",
  device = "pdf",
  width = 10,
  height = 7,
  get.table = TRUE,
  ...
)
}
\arguments{
\item{centrality}{output from \code{quickNet::Centrality}}

\item{prefix}{the prefix of output plot files.}

\item{path}{the path of output files, can be either a relative or absolute path.}

\item{device}{any format supported by \code{ggplot2::ggsave}.}

\item{width}{the width of plot, in inch.}

\item{height}{the height of plot, in inch.}

\item{get.table}{logical. whether get the csv file of centrality. Default is TRUE.}

\item{...}{other parameter from \code{ggplot2::ggsave}.}
}
\value{
a centrality plot will be output to the specified path.
}
\description{
a wrapper function of \code{ggplot2::ggsave}
}
\examples{
data('mtcars')
centrality <- Centrality(EBICglassoNet(mtcars))
get_centrality_plot(centrality, prefix = 'test')

}
