% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_compare_plot.R
\name{get_compare_plot}
\alias{get_compare_plot}
\title{get network difference plot from NetCompare and quickNet results.}
\usage{
get_compare_plot(
  NetCompare,
  network_G,
  maximum = 0.47,
  use.mask = "none",
  output = TRUE,
  prefix = "",
  path = ".",
  device = "pdf",
  width = 10,
  height = 7,
  ...
)
}
\arguments{
\item{NetCompare}{the result of \code{quickNet::NetCompare}.}

\item{network_G}{the result of \code{quickNet::quickNet}. The plot configurations of the results will be the same as network_G, but the legend and pie circle will be removed.}

\item{maximum}{regards the highest of the maximum or highest absolute edge weight as the highest weight to scale the edge widths too. To compare several graphs, set this argument to a higher value than any edge weight in the graphs (typically 1 for correlations). If you want the same scale as network_G, keep it same as the number you set in quickNet.}

\item{use.mask}{only present significant different edges that are non-zero in the EBICglasso networks constructed from data1, data2, or both. \itemize{
\item{"none": present all significantly different edges.}
\item{"net1": only present significantly different edges that are non-zero in the EBICglasso network constructed from data1.}
\item{"net2": only present significantly different edges that are non-zero in the EBICglasso network constructed from data2.}
\item{"both": only present significantly different edges that are both non-zero in the EBICglasso networks constructed from data1 and data2.}
}}

\item{output}{whether output the plots as pdf files.}

\item{prefix}{the prefix of output plot files.}

\item{path}{the path of output files, can be either a relative or absolute path.}

\item{device}{'pdf' or 'svg', deciding the output plot format.}

\item{width}{the width of plot, in inch.}

\item{height}{the height of plot, in inch. \itemize{
\item the parameter prefix, path, width, and height only works when output is TRUE.
}}

\item{...}{... other parameter from \code{pdf} or \code{svg}.}
}
\value{
a list contains the diff_plot, diff_pos_plot and diff_neg_plot,
representing the network difference of global, nw1>nw2 and nw1<nw2.
}
\description{
get network difference plot from NetCompare and quickNet results.
}
\examples{
data('mtcars')
network_G <- quickNet(mtcars)
NetCompare <- NetCompare(mtcars,mtcars^3,it=100)
get_compare_plot(NetCompare, network_G, output = FALSE)

}
