% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lty_matrix.R
\name{get_lty_matrix}
\alias{get_lty_matrix}
\title{Change network linetype based on group/community.}
\usage{
get_lty_matrix(data, group, inner_type = 1, outer_type = 2)
}
\arguments{
\item{data}{The data frame used to construct the data}

\item{group}{The group information defining the community of nodes (must be a list.)}

\item{inner_type}{The line type of inner community edges.}

\item{outer_type}{The line type of outer community edges.}
}
\value{
A matrix that define the linetype of each edge. This can be directly passed to the \code{lty} parameter in function \code{quickNet}
}
\description{
Change network linetype based on group/community.
}
\examples{
data("mtcars")
groups = list(`1` = seq(1,11,by=3),`2` = seq(2,11,by=3),`3` = seq(3,11,by=3))
lty = get_lty_matrix(mtcars, groups)
quickNet(mtcars, groups = groups, lty = lty)


}
