% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_network_plot.R
\name{get_network_plot}
\alias{get_network_plot}
\title{get network plot from qgraph objects.}
\usage{
get_network_plot(
  network,
  prefix = "",
  path = ".",
  device = "pdf",
  width = 10,
  height = 7,
  get.matrix = TRUE,
  ...
)
}
\arguments{
\item{network}{a qgraph object.}

\item{prefix}{the prefix of output plot files.}

\item{path}{the path of output files, can be either a relative or absolute path.}

\item{device}{'pdf' or 'svg', deciding the output plot format.}

\item{width}{the width of plot, in inch.}

\item{height}{the height of plot, in inch.}

\item{get.matrix}{logical. whether get the csv file of graph matrix. Default is TRUE.}

\item{...}{other parameter from \code{pdf} or \code{svg}.}
}
\value{
a network plot will be output to the specified path.
}
\description{
get network plot from qgraph objects.
}
\examples{

data('mtcars')
network <- quickNet(mtcars)
get_network_plot(network, prefix = 'test')

}
