% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stability_plot.R
\name{get_stability_plot}
\alias{get_stability_plot}
\title{get stability plot from stability results.}
\usage{
get_stability_plot(
  stability,
  prefix = "",
  path = ".",
  device = "pdf",
  width = 10,
  height = 7,
  get.table = TRUE,
  ...
)
}
\arguments{
\item{stability}{output from \code{quickNet::Stability}.}

\item{prefix}{the prefix of output plot files.}

\item{path}{the path of output files, can be either a relative or absolute path.}

\item{device}{'pdf' or 'svg', deciding the output plot format.}

\item{width}{the width of plot, in inch.}

\item{height}{the height of plot, in inch.}

\item{get.table}{logical. whether get the csv file of CS-coefficient. Default is TRUE.}

\item{...}{other parameter from \code{pdf} or \code{svg}.}
}
\value{
four plot files will be output to the specified path, if the bridge stability is calculated, the outputs will also contain the bridge stability plot.
}
\description{
get stability plot from stability results.
}
\examples{
data('mtcars')
stability <- Stability(mtcars,nboot = 100)
get_stability_plot(stability,prefix = 'test')

}
