% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalCoeff.R
\name{globalCoeff}
\alias{globalCoeff}
\title{Compute global coefficients of a net work.}
\usage{
globalCoeff(x, list = FALSE)
}
\arguments{
\item{x}{a network produced from \code{quickNet}, or an adjacency matrix of network data.}

\item{list}{whether arrange the results as a list. The default is FALSE.}
}
\value{
\itemize{
\item{\code{globalStrength:} the global strength.}
\item{\code{AGS:} the average of all the edge weights in the network. equals globalStrength/(N * (N-1)/2), where N is the number of nodes in the network.}
\item{\code{ASPL:} the global average shortest path length.}
\item{\code{CC:} the global clustering coefficient. The average clustering coefficient for each node in the network.}
}
}
\description{
Compute global coefficients of a net work.
}
\examples{

data('mtcars')

globalCoeff <- globalCoeff(quickNet(mtcars))

}
