% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netCor.R
\name{netCor}
\alias{netCor}
\title{Calculate the correlation between two networks}
\usage{
netCor(x1, x2, nperm = 1000, graph = TRUE, alternative = "two.sided", ...)
}
\arguments{
\item{x1}{the first network, should be the product of \code{quickNet::quickNet()}, or an adjacency matrix of network data.}

\item{x2}{the second network, should be the product of \code{quickNet::quickNet()}, or an adjacency matrix of network data.}

\item{nperm}{the number of times to permute the data.}

\item{graph}{a logical indicating whether to produce a summary graph.}

\item{alternative}{a character string defining the alternative hypothesis: "two.sided" (default), "less", "greater", or any unambiguous abbreviation of these.}

\item{...}{further arguments to be passed to plot() (to add a title, change the axis labels, and so on).}
}
\value{
\itemize{
\item{\code{z.stat:} 	the Z-statistic (sum of rows*columns of lower triangle) of the data matrices.}
\item{\code{p:} P-value (quantile of the observed Z-statistic in the permutation distribution).}
\item{\code{alternative:} the alternative hypothesis.}
}
}
\description{
This function computes Mantel's permutation test for similarity of two matrices. It permutes the rows and columns of the second matrix randomly and calculates a Z-statistic.
}
\details{
The function calculates a Z-statistic for the Mantel test, equal to the sum of the pairwise product of the lower triangles of the permuted matrices, for each permutation of rows and columns. It compares the permuted distribution with the Z-statistic observed for the actual data.
}
\examples{
data('mtcars')
net1 <- quickNet(mtcars)
net2 <- quickNet(mtcars^3)
netCor(net1, net2)

}
